backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-socketio==5.9.0
PyMySQL==1.1.0
mysql-connector-python
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==2.3.7
cryptography==41.0.4
Pillow==10.0.1
